python-otel/pyproject.toml
[tool.poetry]
name = "otelp"
version = "0.1.0"
description = ""
authors = ["Hua Liang <huliang@microsoft.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
opentelemetry-sdk = "^1.22.0"
opentelemetry-api = "^1.22.0"
opentelemetry-exporter-otlp = "^1.22.0"
opentelemetry-instrumentation-requests = "^0.45b0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
